requirements.txt
transformers>=4.35.0
torch>=2.0.0
gradio>=4.0.0
pillow>=10.0.0